--- src/pages/ProductList/component/Button/Button.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import './Button.scss';
-
-const Button = () => {
-  return (
-    <div className="Button">
-      <button>1</button>
-      <button>2</button>
-      <button>3</button>
-    </div>
-  );
-};
-
-export default Button;
